Add/rename cars
Add all missing cars, rename existing cars to match current iRacing names, sort names in file to match iRacing order. All missing car values are pulled from the iRacing API.

Note that the sort order is not the one used in SoundShift; this is because the sorting algorithm used in the original SoundShift code does not correctly sort non-alphabetical characters. It is possible to modify the executable yourself to change the sort option.